Extract resolveFunction() :- Phase 3
