+++ compiler/resolution/resolveFunction.cpp
@@ -27,8 +27,10 @@
 #include "expr.h"
 #include "ForLoop.h"
 #include "ForallStmt.h"
+#include "iterator.h"
 #include "LoopStmt.h"
 #include "ParamForLoop.h"
+#include "passes.h"
 #include "resolution.h"
 #include "resolveIntents.h"
 #include "stmt.h"
@@ -47,6 +49,373 @@ static void fixTypeNames(AggregateType* at);
 static void insertUnrefForArrayReturn(FnSymbol* fn);
 
 static void instantiateDefaultConstructor(FnSymbol* fn);
+
+/************************************* | **************************************
+*                                                                             *
+*                                                                             *
+*                                                                             *
+************************************** | *************************************/
+
+void resolveFormalsAndFunction(FnSymbol* fn) {
+  resolveFormals(fn);
+  resolveFunction(fn);
+}
+
+/************************************* | **************************************
+*                                                                             *
+*                                                                             *
+*                                                                             *
+************************************** | *************************************/
+
+static bool      convertAtomicFormalTypeToRef(ArgSymbol* formal,
+                                              FnSymbol*  fn);
+
+static void      resolveSpecifiedReturnType(FnSymbol* fn);
+
+static void      protoIteratorClass(FnSymbol* fn);
+
+static FnSymbol* protoIteratorMethod(IteratorInfo* ii,
+                                     const char*   name,
+                                     Type*         retType);
+
+void resolveFormals(FnSymbol* fn) {
+  static Vec<FnSymbol*> done;
+
+  if (!fn->hasFlag(FLAG_GENERIC)) {
+    if (done.set_in(fn))
+      return;
+
+    done.set_add(fn);
+
+    for_formals(formal, fn) {
+      if (formal->type == dtUnknown) {
+        if (!formal->typeExpr) {
+          formal->type = dtObject;
+
+        } else {
+          resolveBlockStmt(formal->typeExpr);
+
+          formal->type = formal->typeExpr->body.tail->getValType();
+        }
+      }
+
+      //
+      // Fix up value types that need to be ref types.
+      //
+      if (formal->type->symbol->hasFlag(FLAG_REF))
+        // Already a ref type, so done.
+        continue;
+
+      // Don't pass dtString params in by reference
+      if(formal->type == dtString && formal->hasFlag(FLAG_INSTANTIATED_PARAM))
+        continue;
+
+      if (formal->type->symbol->hasFlag(FLAG_RANGE) &&
+          (formal->intent == INTENT_BLANK || formal->intent == INTENT_IN) &&
+          fn->hasFlag(FLAG_BEGIN)) {
+        // For begin functions, copy ranges in if passed by blank intent.
+        // This is a temporary workaround.
+        // * arguably blank intent for ranges should be 'const ref',
+        //   but the current compiler uses a mix of 'const in' and 'const ref'.
+        // * resolveIntents is changing INTENT_IN to INTENT_REF
+        //   which interferes with the logic in parallel.cpp
+        //   (another approach to that problem might be to separately
+        //    store the 'requested intent' and the 'concrete intent').
+        formal->intent = INTENT_CONST_IN;
+      }
+
+      if (formal->intent == INTENT_INOUT ||
+          formal->intent == INTENT_OUT ||
+          formal->intent == INTENT_REF ||
+          formal->intent == INTENT_CONST_REF ||
+          convertAtomicFormalTypeToRef(formal, fn) ||
+          formal->hasFlag(FLAG_WRAP_WRITTEN_FORMAL) ||
+          (formal == fn->_this &&
+           !formal->hasFlag(FLAG_TYPE_VARIABLE) &&
+           (isUnion(formal->type) ||
+            isRecord(formal->type)))) {
+        makeRefType(formal->type);
+
+        if (formal->type->refType) {
+          formal->type = formal->type->refType;
+          // The type of the formal is its own ref type!
+
+        } else {
+          formal->qual = QUAL_REF;
+        }
+      }
+
+      if (isRecordWrappedType(formal->type) &&
+          fn->hasFlag(FLAG_ITERATOR_FN)) {
+        // Pass domains, arrays into iterators by ref.
+        // This is a temporary workaround for issues with
+        // iterator lowering.
+        // It is temporary because we expect more of the
+        // compiler to handle 'refness' of an ArgSymbol
+        // in the future.
+        makeRefType(formal->type);
+
+        formal->type = formal->type->refType;
+      }
+
+      // Adjust tuples for intent.
+      // This should not apply to 'ref' , 'out', or 'inout' formals,
+      // but these are currently turned into reference types above.
+      // It probably should not apply to 'const ref' either but
+      // that is more debatable.
+      if (formal->type->symbol->hasFlag(FLAG_TUPLE) &&
+          !formal->hasFlag(FLAG_TYPE_VARIABLE) &&
+          !(formal == fn->_this) &&
+          !doNotChangeTupleTypeRefLevel(fn, false)) {
+
+        // Let 'in' intent work similarly to the blank intent.
+        IntentTag intent = formal->intent;
+        if (intent == INTENT_IN) intent = INTENT_BLANK;
+        AggregateType* tupleType = toAggregateType(formal->type);
+        INT_ASSERT(tupleType);
+        Type* newType = computeTupleWithIntent(intent, tupleType);
+        formal->type = newType;
+      }
+
+    }
+
+    if (fn->retExprType) {
+      resolveSpecifiedReturnType(fn);
+    }
+
+    resolvedFormals.set_add(fn);
+  }
+}
+
+//
+// Generally, atomics must also be passed by reference when
+// passed by blank intent.  The following expression checks for
+// these cases by looking for atomics passed by blank intent and
+// changing their type to a ref type.  Interestingly, this
+// conversion does not seem to be required for single-locale
+// compilation, but it is for multi-locale.  Otherwise, updates
+// to atomics are lost (as demonstrated by
+// test/functions/bradc/intents/test_pass_atomic.chpl).
+//
+// I say "generally" because there are a few cases where passing
+// atomics by reference breaks things -- primarily in
+// constructors, assignment operators, and tuple construction.
+// So we have some unfortunate special checks that dance around
+// these cases.
+//
+// While I can't explain precisely why these special cases are
+// required yet, here are the tests that tend to have problems
+// without these special conditions:
+//
+//   test/release/examples/benchmarks/hpcc/ra-atomics.chpl
+//   test/types/atomic/sungeun/no_atomic_assign.chpl
+//   test/functions/bradc/intents/test_construct_atomic_intent.chpl
+//   test/users/vass/barrierWF.test-1.chpl
+//   test/studies/shootout/spectral-norm/spectralnorm.chpl
+//   test/release/examples/benchmarks/ssca2/SSCA2_main.chpl
+//   test/parallel/taskPar/sungeun/barrier/*.chpl
+//
+static bool convertAtomicFormalTypeToRef(ArgSymbol* formal, FnSymbol* fn) {
+  return (formal->intent == INTENT_BLANK &&
+          !formal->hasFlag(FLAG_TYPE_VARIABLE) &&
+          isAtomicType(formal->type))
+    && fn->name != astrSequals
+    && !fn->hasFlag(FLAG_DEFAULT_CONSTRUCTOR)
+    && !fn->hasFlag(FLAG_CONSTRUCTOR)
+    && !fn->hasFlag(FLAG_BUILD_TUPLE);
+}
+
+
+static void resolveSpecifiedReturnType(FnSymbol* fn) {
+  Type* retType;
+
+  resolveBlockStmt(fn->retExprType);
+
+  retType = fn->retExprType->body.tail->typeInfo();
+
+  if (SymExpr* se = toSymExpr(fn->retExprType->body.tail)) {
+    // Try resolving global type aliases
+    if (se->symbol()->hasFlag(FLAG_TYPE_VARIABLE))
+      retType = resolveTypeAlias(se);
+
+    if (retType->symbol->hasFlag(FLAG_GENERIC)) {
+      SET_LINENO(fn->retExprType->body.tail);
+
+      // Try resolving records/classes with default types
+      // e.g. range
+      retType = resolveDefaultGenericTypeSymExpr(se);
+    }
+  }
+
+  fn->retType = retType;
+
+  if (retType != dtUnknown) {
+    // Difficult to call e.g. _unref_type in build/normalize
+    // b/c of bugs in pulling out function symbols.
+    // So we make sure returned tuple types capture values here.
+    if (retType->symbol->hasFlag(FLAG_TUPLE)   == true   &&
+        doNotChangeTupleTypeRefLevel(fn, true) == false) {
+      AggregateType* tupleType = toAggregateType(retType);
+
+      INT_ASSERT(tupleType);
+
+      retType = getReturnedTupleType(fn, tupleType);
+
+      fn->retType = retType;
+
+    } else if (fn->returnsRefOrConstRef()) {
+      makeRefType(retType);
+
+      retType     = fn->retType->refType;
+      fn->retType = retType;
+    }
+
+    fn->retExprType->remove();
+
+    if (fn->isIterator() && !fn->iteratorInfo) {
+      // Note: protoIteratorClass changes fn->retType
+      // to the iterator record. The original return type
+      // is stored here in retType.
+      protoIteratorClass(fn);
+    }
+  }
+
+  // Also update the return symbol type
+  Symbol* ret = fn->getReturnSymbol();
+
+  if (ret->type == dtUnknown) {
+    // uses the local variable saving the resolved declared return type
+    // since for iterators, fn->retType is the iterator record.
+    ret->type = retType;
+  }
+}
+
+static void protoIteratorClass(FnSymbol* fn) {
+  INT_ASSERT(!fn->iteratorInfo);
+
+  SET_LINENO(fn);
+
+  IteratorInfo* ii = new IteratorInfo();
+  fn->iteratorInfo = ii;
+  fn->iteratorInfo->iterator = fn;
+
+  const char* className = astr(fn->name);
+  if (fn->_this)
+    className = astr(className, "_", fn->_this->type->symbol->cname);
+
+  ii->iclass = new AggregateType(AGGREGATE_CLASS);
+
+  TypeSymbol* cts = new TypeSymbol(astr("_ic_", className), ii->iclass);
+
+  cts->addFlag(FLAG_ITERATOR_CLASS);
+  cts->addFlag(FLAG_POD);
+
+  ii->iclass->addRootType();
+
+  fn->defPoint->insertBefore(new DefExpr(cts));
+
+  ii->irecord = new AggregateType(AGGREGATE_RECORD);
+
+  TypeSymbol* rts = new TypeSymbol(astr("_ir_", className), ii->irecord);
+
+  rts->addFlag(FLAG_ITERATOR_RECORD);
+
+  // TODO -- do a better job of deciding if an iterator record is
+  // POD or not POD.
+  rts->addFlag(FLAG_NOT_POD);
+
+  if (fn->retTag == RET_REF) { // TODO -- handle RET_CONST_REF
+    rts->addFlag(FLAG_REF_ITERATOR_CLASS);
+  }
+
+  fn->defPoint->insertBefore(new DefExpr(rts));
+
+  ii->tag      = it_iterator;
+  ii->advance  = protoIteratorMethod(ii, "advance",  dtVoid);
+  ii->zip1     = protoIteratorMethod(ii, "zip1",     dtVoid);
+  ii->zip2     = protoIteratorMethod(ii, "zip2",     dtVoid);
+  ii->zip3     = protoIteratorMethod(ii, "zip3",     dtVoid);
+  ii->zip4     = protoIteratorMethod(ii, "zip4",     dtVoid);
+  ii->hasMore  = protoIteratorMethod(ii, "hasMore",  dtInt[INT_SIZE_DEFAULT]);
+  ii->getValue = protoIteratorMethod(ii, "getValue", fn->retType);
+  ii->init     = protoIteratorMethod(ii, "init",     dtVoid);
+  ii->incr     = protoIteratorMethod(ii, "incr",     dtVoid);
+
+  // Save the iterator info in the iterator record.
+  // The iterator info is still owned by the iterator function.
+  ii->irecord->iteratorInfo        = ii;
+  ii->irecord->scalarPromotionType = fn->retType;
+
+  fn->retType = ii->irecord;
+  fn->retTag  = RET_VALUE;
+
+  makeRefType(fn->retType);
+
+  ii->getIterator = new FnSymbol("_getIterator");
+
+  ii->getIterator->addFlag(FLAG_AUTO_II);
+  ii->getIterator->addFlag(FLAG_INLINE);
+  ii->getIterator->retType = ii->iclass;
+
+  ii->getIterator->insertFormalAtTail(new ArgSymbol(INTENT_BLANK,
+                                                    "ir",
+                                                    ii->irecord));
+
+  VarSymbol* ret = newTemp("_ic_", ii->iclass);
+
+  ii->getIterator->insertAtTail(new DefExpr(ret));
+
+  CallExpr* icAllocCall = callChplHereAlloc(ret->typeInfo());
+
+  ii->getIterator->insertAtTail(new CallExpr(PRIM_MOVE, ret, icAllocCall));
+  ii->getIterator->insertAtTail(new CallExpr(PRIM_SETCID, ret));
+  ii->getIterator->insertAtTail(new CallExpr(PRIM_RETURN, ret));
+
+  fn->defPoint->insertBefore(new DefExpr(ii->getIterator));
+
+  // Save the iterator info in the iterator class also.
+  // This makes it easy to obtain the iterator given
+  // just a symbol of the iterator class type.  This may include _getIterator
+  // and _getIteratorZip functions in the module code.
+  ii->iclass->iteratorInfo = ii;
+  normalize(ii->getIterator);
+  resolveFunction(ii->getIterator);  // No shortcuts.
+}
+
+static FnSymbol* protoIteratorMethod(IteratorInfo* ii,
+                                     const char*   name,
+                                     Type*         retType) {
+  FnSymbol* fn = new FnSymbol(name);
+
+  fn->addFlag(FLAG_AUTO_II);
+
+  if (strcmp(name, "advance") != 0) {
+    fn->addFlag(FLAG_INLINE);
+  }
+
+  fn->insertFormalAtTail(new ArgSymbol(INTENT_BLANK, "_mt", dtMethodToken));
+
+  fn->addFlag(FLAG_METHOD);
+
+  fn->_this = new ArgSymbol(INTENT_BLANK, "this", ii->iclass);
+
+  fn->_this->addFlag(FLAG_ARG_THIS);
+
+  fn->retType = retType;
+
+  fn->insertFormalAtTail(fn->_this);
+
+  ii->iterator->defPoint->insertBefore(new DefExpr(fn));
+
+  normalize(fn);
+
+  // Pretend that this function is already resolved.
+  // Its body will be filled in during the lowerIterators pass.
+  fn->addFlag(FLAG_RESOLVED);
+
+  return fn;
+}
+
 
 /************************************* | **************************************
 *                                                                             *
@@ -349,8 +718,8 @@ static void insertUnrefForArrayReturn(FnSymbol* fn) {
         // TODO: Should we check if the RHS is a symbol with 'no auto
         // destroy' on it? If it is, then we'd be copying the RHS and it
         // would never be destroyed...
-        if (rhsType->symbol->hasFlag(FLAG_ARRAY) &&
-            isTypeExpr(call->get(2)) == false) {
+        if (rhsType->symbol->hasFlag(FLAG_ARRAY) == true &&
+            isTypeExpr(call->get(2))             == false) {
           Expr*      origRHS = call->get(2)->remove();
           VarSymbol* tmp     = newTemp(arrayUnrefName, rhsType);
 
@@ -382,7 +751,9 @@ static void insertUnrefForArrayReturn(FnSymbol* fn) {
 
             // Remove now-useless AST
             tmp->defPoint->remove();
+
             init_unref_tmp->remove();
+
             INT_ASSERT(unrefCall->inTree() == false);
           }
         }
@@ -687,10 +1058,10 @@ static bool formalRequiresTemp(ArgSymbol* formal) {
 // passing an argument of type 't'.
 //
 static bool backendRequiresCopyForIn(Type* t) {
-  return (isRecord(t) ||
-          isUnion(t) ||
-          t->symbol->hasFlag(FLAG_ARRAY) ||
-          t->symbol->hasFlag(FLAG_DOMAIN));
+  return isRecord(t)                     == true ||
+         isUnion(t)                      == true ||
+         t->symbol->hasFlag(FLAG_ARRAY)  == true ||
+         t->symbol->hasFlag(FLAG_DOMAIN) == true;
 }
 
 
@@ -711,7 +1082,7 @@ static void addLocalCopiesAndWritebacks(FnSymbol*  fn,
   // Enumerate the formals that have local temps.
   form_Map(SymbolMapElem, e, formals2vars) {
     ArgSymbol* formal = toArgSymbol(e->key); // Get the formal.
-    Symbol*    tmp    = e->value; // Get the temp.
+    Symbol*    tmp    = e->value;            // Get the temp.
 
     SET_LINENO(formal);
